+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.1.0
 
-* This package helps to encryption and decryption with or without secret key.
+* Decode JWT token and expiration date time.
 
 ## 1.0.0